+++ lib/src/main/java/org/eclipse/tsp/java/client/api/experiment/ExperimentApi.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 import java.util.UUID;
 
 import org.eclipse.tsp.java.client.api.AbstractTspApi;
+import org.eclipse.tsp.java.client.core.restclient.RestClient;
+import org.eclipse.tsp.java.client.core.tspclient.TspClientResponse;
 import org.eclipse.tsp.java.client.shared.query.Query;
-import org.eclipse.tsp.java.client.shared.restclient.RestClient;
-import org.eclipse.tsp.java.client.shared.tspclient.TspClientResponse;
 
 public class ExperimentApi extends AbstractTspApi {
     final String EXPERIMENT_API_URL = "%s/experiments";
