+++ lib/src/main/java/org/eclipse/tsp/java/client/api/outputdescriptor/OutputDescriptorApi.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import java.util.UUID;
 
 import org.eclipse.tsp.java.client.api.AbstractTspApi;
-import org.eclipse.tsp.java.client.shared.restclient.RestClient;
-import org.eclipse.tsp.java.client.shared.tspclient.TspClientResponse;
+import org.eclipse.tsp.java.client.core.restclient.RestClient;
+import org.eclipse.tsp.java.client.core.tspclient.TspClientResponse;
 
 public class OutputDescriptorApi extends AbstractTspApi {
     private final String OUTPUT_DESCRIPTOR_API_URL = "%s/experiments/%s/outputs";
