+++ lib/src/main/java/org/eclipse/tsp/java/client/api/trace/TraceApi.java
@@ -6,9 +6,9 @@ import java.util.Optional;
 import java.util.UUID;
 
 import org.eclipse.tsp.java.client.api.AbstractTspApi;
+import org.eclipse.tsp.java.client.core.restclient.RestClient;
+import org.eclipse.tsp.java.client.core.tspclient.TspClientResponse;
 import org.eclipse.tsp.java.client.shared.query.Query;
-import org.eclipse.tsp.java.client.shared.restclient.RestClient;
-import org.eclipse.tsp.java.client.shared.tspclient.TspClientResponse;
 
 public class TraceApi extends AbstractTspApi {
     private final String TRACE_API_URL = "%s/traces";
