+++ lib/src/main/java/org/eclipse/tsp/java/client/core/restclient/PclConnectionStatus.java
@@ -1,4 +1,4 @@
-package org.eclipse.tsp.java.client.shared.restclient;
+package org.eclipse.tsp.java.client.core.restclient;
 
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
