+++ lib/src/test/java/org/eclipse/tsp/java/client/protocol/restclient/ConnectionStatusTest.java
@@ -2,8 +2,8 @@ package org.eclipse.tsp.java.client.protocol.restclient;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.eclipse.tsp.java.client.shared.restclient.ConnectionStatus;
-import org.eclipse.tsp.java.client.shared.restclient.PclConnectionStatus;
+import org.eclipse.tsp.java.client.core.restclient.ConnectionStatus;
+import org.eclipse.tsp.java.client.core.restclient.PclConnectionStatus;
 import org.junit.jupiter.api.Test;
 
 public class ConnectionStatusTest {
